fix: Deleted all messed up code
