+++ app/src/main/java/com/example/wellfed/ActivityBase.java
@@ -16,4 +16,8 @@ public abstract class ActivityBase extends AppCompatActivity {
         }
         return super.onOptionsItemSelected(item);
     }
+
+    public abstract void deleteIngredient();
+
+    public abstract void delete();
 }
