+++ app/src/main/java/com/example/wellfed/navigation/NavigationCollectionAdapter.java
@@ -7,7 +7,6 @@ import androidx.viewpager2.adapter.FragmentStateAdapter;
 
 import com.example.wellfed.recipe.RecipeBookFragment;
 
-import com.example.wellfed.ingredient.IngredientStorageFragment;
 import com.example.wellfed.mealplan.MealBookFragment;
 import com.example.wellfed.shoppingcart.ShoppingCartFragment;
